Add files via upload
Fashion Image Classification
Fashion-MNIST is a dataset of Zalando's article images—consisting of a training set of 60,000 examples and a test set of 10,000 examples. Each example is a 28x28 grayscale image, associated with a label from 10 classes.
I Built a classifier for the Fashion-MNIST dataset that achieves over 85% accuracy on the test set.